Fix of push Objects - Pre Build of all APIs 3

+++ src/models/mongo/store/painting-store.js
@@ -5,20 +5,19 @@ export const paintingMongoStore = {
         return await Painting.find().lean() || [];
     },
 
+    async getAllPaintingsByGallery(gallery) {
+      return await Painting.find({ gallery: gallery }).lean() || [];
+    },
+
+    async getAllPaintingsByArtist(artist) {
+      return await Painting.find({ artist: artist }).lean() || [];
+    },
+
+    async getAllPaintingsByEpoch(epoch) {
+        return await Painting.find({ epoch: epoch }).lean() || [];
+    },
     async getPaintingById(id) {
         return await Painting.findOne({ _id: id }) || null;
-    },
-
-    async getPaintingsByEpoch(epoch) {
-        return await Painting.find({ epoch: epoch }).lean() || null;
-    },
-
-    async getPaintingsByArtist(artist) {
-        return await Painting.find({ artist: artist }).lean() || null;
-    },
-
-    async getPaintingsByGallery(gallery) {
-        return await Painting.find({ gallery: gallery }).lean() || null;
     },
 
     /**
@@ -27,15 +26,51 @@ export const paintingMongoStore = {
     * @param {*} paintingCreated
     */
      async createPainting(paintingCreated) {
+       console.log(`handed over: \n${Object.values(paintingCreated)}`);
         const alreadyCreated = await Painting.findOne({
             title: paintingCreated.title,
-            artist: paintingCreated.artist,
+            epoch: paintingCreated.epoch._id,
+            artist: paintingCreated.artist._id,
+            gallery: paintingCreated.gallery._id,
         });
+        console.log(alreadyCreated);
         if (alreadyCreated) return alreadyCreated;
         return await new Painting(paintingCreated).save();
     },
 
-    async deleteAll() {
-      await Painting.deleteMany({});
-    },
+    /**
+  * The deleteMany() returns a document containing the deleteCount field
+  * that stores the number of deleted documents.
+  * @returns {Number}
+  * - n >= 0 for successful deletion
+  * - -1 for missing rights
+  */
+   async deleteAll(user) {
+     console.log("deletAll");
+   if (user.rank > 0) {
+     return await Painting.deleteMany({});
+   } return -1;
+ },
+
+ /**
+ * This method deletes an entry of the database with the given rank
+ * @param {String} deletionPaintingId
+ * @param {*} user
+ * @returns
+ * - 1 for successful deletion
+ * - 0 for no possible entry
+ * - -1 for missing rights
+ */
+ async deletePaintingById(deletionPaintingId, user) {
+   try {
+     const painting = await this.getPaintingById({ _id: deletionPaintingId });
+     if (String(user._id) === String(painting.user) || user.rank > 0) {
+       await Painting.deleteOne({ _id: deletionPaintingId });
+       return 1;
+     }
+     return -1;
+   } catch (error) {
+     return 0;
+   }
+ },
 };
